requirements.txt
numpy
matplotlib
pygame
torch
torchvision
torchaudio
PyYAML